+++ neural_compressor/pruner/README.md
@@ -122,36 +122,33 @@ Users can pass the customized training/evaluation functions to `Pruning` in vari
 
 
 
-The following section exemplifies how to use hooks in user pass-in training function to perform model pruning. Through the pruning API, multiple pruner objects are supported in one single Pruning object to enable layer-specific configurations and a default setting is used as a complement.  
-
-
-
+The following section exemplifies how to use hooks in user pass-in training function to perform model pruning. Through the pruning API, multiple pruner objects are supported in one single Pruning object to enable layer-specific configurations and a default setting is used as a complement.
 
 ```python
-from neural_compressor.pruning import Pruning, WeightPruningConfig
+from neural_compressor.pruner.pruning import Pruning, WeightPruningConfig
 
 config = WeightPruningConfig(
     pruning_configs,  # An example of pruning_configs is shown below.
-    target_sparsity=0.8, start_step=1, end_step=10, pruning_frequency=1 # Default pruning setting.
+    target_sparsity=0.8, start_step=1, end_step=10, pruning_frequency=1  # Default pruning setting.
 )
 prune = Pruning(config)  # Pruning constructor.
-prune.model = model      # Set model object to prune.
-prune.on_train_begin()   # Execute on_train_begin hook before training.
+prune.model = model  # Set model object to prune.
+prune.on_train_begin()  # Execute on_train_begin hook before training.
 for epoch in range(num_train_epochs):
-    model.train()    
-    prune.on_epoch_begin(epoch)    # Execute on_epoch_begin hook before each epoch.
+    model.train()
+    prune.on_epoch_begin(epoch)  # Execute on_epoch_begin hook before each epoch.
     for step, batch in enumerate(train_dataloader):
-        prune.on_step_begin(step)  # Execute on_step_begin hook before each step.
+            prune.on_step_begin(step)  # Execute on_step_begin hook before each step.
         outputs = model(**batch)
         loss = outputs.loss
         loss.backward()
-        prune.on_before_optimizer_step()  #Execute on_before_optimizer_step() hook before optimization.
+        prune.on_before_optimizer_step()  # Execute on_before_optimizer_step() hook before optimization.
         optimizer.step()
-        prune.on_after_optimizer_step()   #Execute on_after_optimizer_step() hook after optimization.
-        scheduler.step()  # Update learning rate schedule
+prune.on_after_optimizer_step()  # Execute on_after_optimizer_step() hook after optimization.
+        scheduler.step()   # Update learning rate schedule
         model.zero_grad()
         prune.on_step_end()  # Execute on_step_end hook after each step.
-    prune.on_epoch_end()  # Execute on_epoch_end hook after each epoch.
+prune.on_epoch_end()  # Execute on_epoch_end hook after each epoch.
 ...
 ```
