+++ test/pruning/test_pruning_config_new.py
@@ -1,5 +1,3 @@
-import os
-import shutil
 import unittest
 
 import torch
@@ -8,7 +6,7 @@ import torch.nn as nn
 
 from neural_compressor.data import Datasets
 from neural_compressor.experimental.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
-from neural_compressor.pruning import WeightPruningConfig
+from neural_compressor.pruner.pruning import WeightPruningConfig
 #from neural_compressor.pruning import Pruning # pytorch_pruner
 from neural_compressor.experimental.pruning_v2 import Pruning
